+++ react/features/reactions/functions.web.ts
@@ -1,5 +1,4 @@
 import { IReduxState } from '../app/types';
-import { getToolbarButtons } from '../base/config/functions.web';
 
 import { shouldDisplayReactionsButtons } from './functions.any';
 
@@ -22,5 +21,7 @@ export function getReactionsMenuVisibility(state: IReduxState): boolean {
  * @returns {boolean}
  */
 export function isReactionsButtonEnabled(state: IReduxState) {
-    return Boolean(getToolbarButtons(state).includes('reactions')) && shouldDisplayReactionsButtons(state);
+    const { toolbarButtons } = state['features/toolbox'];
+
+    return Boolean(toolbarButtons?.includes('reactions')) && shouldDisplayReactionsButtons(state);
 }
