+++ react/features/recording/components/LiveStream/web/LiveStreamButton.ts
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux';
 
 import { IReduxState } from '../../../../app/types';
-import { getToolbarButtons } from '../../../../base/config/functions.web';
 import { openDialog } from '../../../../base/dialog/actions';
 import { translate } from '../../../../base/i18n/functions';
 import AbstractLiveStreamButton, {
@@ -50,11 +49,11 @@ class LiveStreamButton extends AbstractLiveStreamButton<IProps> {
  */
 function _mapStateToProps(state: IReduxState, ownProps: any) {
     const abstractProps = _abstractMapStateToProps(state, ownProps);
-    const toolbarButtons = getToolbarButtons(state);
+    const { toolbarButtons } = state['features/toolbox'];
     let { visible } = ownProps;
 
     if (typeof visible === 'undefined') {
-        visible = toolbarButtons.includes('livestreaming') && abstractProps.visible;
+        visible = Boolean(toolbarButtons?.includes('livestreaming') && abstractProps.visible);
     }
 
     return {
