+++ react/features/toolbox/types.ts
@@ -6,3 +6,41 @@ export interface IToolboxButton {
     group: number;
     key: string;
 }
+
+export type ToolbarButton = 'camera' |
+    'chat' |
+    'closedcaptions' |
+    'desktop' |
+    'download' |
+    'embedmeeting' |
+    'etherpad' |
+    'feedback' |
+    'filmstrip' |
+    'fullscreen' |
+    'hangup' |
+    'help' |
+    'highlight' |
+    'invite' |
+    'linktosalesforce' |
+    'livestreaming' |
+    'microphone' |
+    'mute-everyone' |
+    'mute-video-everyone' |
+    'noisesuppression' |
+    'participants-pane' |
+    'profile' |
+    'raisehand' |
+    'reactions' |
+    'recording' |
+    'security' |
+    'select-background' |
+    'settings' |
+    'shareaudio' |
+    'sharedvideo' |
+    'shortcuts' |
+    'stats' |
+    'tileview' |
+    'toggle-camera' |
+    'videoquality' |
+    'whiteboard' |
+    '__end';
